apps/ai-worker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
openai==1.3.7
anthropic==0.7.8
tiktoken==0.5.2
structlog==23.2.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
vosk==0.3.45
langdetect==1.0.9
bcrypt==4.1.2
